Break actions button into separate file.
